+++ src/nvlib/exporter/nv_doc_exporter.py
@@ -6,7 +6,6 @@ License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 from datetime import datetime
 import os
-from tkinter import messagebox
 
 from novxlib.converter.export_target_factory import ExportTargetFactory
 from novxlib.file.doc_open import open_document
@@ -63,8 +62,9 @@ class NvDocExporter:
         OdtWXref,
         ]
 
-    def __init__(self):
+    def __init__(self, ui):
         """Create strategy class instances."""
+        self._ui = ui
         self.exportTargetFactory = ExportTargetFactory(self.EXPORT_TARGET_CLASSES)
         self._source = None
         self._target = None
@@ -133,9 +133,9 @@ class NvDocExporter:
         self._targetFileDate = datetime.now().replace(microsecond=0).isoformat(sep=' ')
         if kwargs.get('show', True):
             askOpen = kwargs.get('ask', True) and prefs['ask_doc_open']
-            if not askOpen or messagebox.askyesno(
+            if not askOpen or self._ui.ask_yes_no(
+                _('{} created.\n\nOpen now?').format(norm_path(self._target.DESCRIPTION)),
                 title=self._target.novel.title,
-                message=_('{} created.\n\nOpen now?').format(norm_path(self._target.DESCRIPTION))
                 ):
                 open_document(self._target.filePath)
         return _('Created {0} on {1}.').format(self._target.DESCRIPTION, self._targetFileDate)